Move data portion of `SemanticModel` to dbt/artifacts
