+++ .changes/unreleased/Under the Hood-20240129-163800.yaml
@@ -1,0 +1,6 @@
+kind: Under the Hood
+body: Move data portion of `SemanticModel` to dbt/artifacts
+time: 2024-01-29T16:38:00.245253-08:00
+custom:
+  Author: QMalcolm
+  Issue: "9387"
