+++ core/dbt/artifacts/resources/__init__.py
@@ -41,5 +41,6 @@ from dbt.artifacts.resources.v1.semantic_model import (
     MeasureAggregationParameters,
     NodeRelation,
     NonAdditiveDimension,
+    SemanticModel,
     SemanticModelConfig,
 )
